requirements.txt
streamlit>=1.37.0
langchain>=0.2.10
langchain-openai>=0.1.21
openai>=1.51.0
pdfplumber>=0.11.0
python-dotenv>=1.0.1
